docs/RAG/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.5
bleach==6.2.0
blockbuster==1.5.25
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
cohere==5.20.0
comm==0.2.3
cryptography==44.0.3
dataclasses-json==0.6.7
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
executing==2.2.1
fastapi==0.116.1
fastavro==1.12.1
fastjsonschema==2.21.2
filelock==3.20.0
forbiddenfruit==0.1.4
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==1.0.1
idna==3.10
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jsonschema_rs==0.29.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
langchain==0.3.27
langchain-cohere==0.4.6
langchain-community==0.3.21
langchain-core==0.3.79
langchain-experimental==0.3.4
langchain-mcp-adapters==0.1.10
langchain-openai==0.3.33
langchain-text-splitters==0.3.9
langgraph==0.6.7
langgraph-api==0.4.20
langgraph-checkpoint==2.1.1
langgraph-checkpoint-postgres==2.0.23
langgraph-cli==0.4.2
langgraph-prebuilt==0.6.4
langgraph-runtime-inmem==0.12.0
langgraph-sdk==0.2.6
langsmith==0.3.45
lark==1.2.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.15.0
mdurl==0.1.2
mistune==3.1.4
multidict==6.7.0
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook_shim==0.2.4
numpy==2.3.4
openai==1.107.2
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
platformdirs==4.4.0
pluggy==1.6.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.0.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg-pool==3.2.6
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.1
Send2Trash==1.8.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.47.3
structlog==25.4.0
tenacity==9.1.2
terminado==0.18.1
tiktoken==0.11.0
tinycss2==1.4.0
tokenizers==0.22.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
truststore==0.10.4
typer==0.19.2
typer-slim==0.20.0
types-python-dateutil==2.9.0.20250822
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
xxhash==3.5.0
yarl==1.22.0
zstandard==0.23.0
